refactor: pages structure based on conventions
