+++ resources/ts/pages/Admin/resources/index.tsx
