+++ resources/ts/pages/Admin/settings/contracts/index.tsx
