+++ resources/ts/pages/Admin/settings/resource-types/index.tsx
